docker_deploy/web-app/requirements.txt
Django>3.0
psycopg2
djangorestframework==3.14.0
